refact : 각 도메인 커스텀 예외 클래스들 정리
