--- src/main/java/co/kr/jurumarble/exception/comment/CommentExceptionHandler.java
@@ -1,22 +1,0 @@
-package co.kr.jurumarble.exception.comment;
-
-
-import co.kr.jurumarble.exception.ExceptionMessage;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-@Slf4j
-@RequiredArgsConstructor
-public class CommentExceptionHandler {
-
-    @ExceptionHandler(CommentNotFoundException.class)
-    public ResponseEntity<ExceptionMessage> handle(CommentNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(ExceptionMessage.of(e.getStatus(), e.getMessage()));
-    }
-}
