+++ src/main/java/co/kr/jurumarble/exception/common/CustomExceptionHandler.java
@@ -1,20 +1,30 @@
-package co.kr.jurumarble.exception.token;
+package co.kr.jurumarble.exception.common;
 
 import co.kr.jurumarble.exception.ExceptionMessage;
 import co.kr.jurumarble.exception.StatusEnum;
+import co.kr.jurumarble.exception.token.TokenNotFoundException;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
-@Slf4j
-@RequiredArgsConstructor
-public class TokenExceptionHandler {
+public class CustomExceptionHandler {
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity<ExceptionMessage> handle(CustomException e) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                .body(ExceptionMessage.of(e.getStatus(), e.getMessage()));
+    }
+
+    @ExceptionHandler(BindException.class)
+    public ResponseEntity<ExceptionMessage> bindException(BindException e) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                .body(ExceptionMessage.of(e.getBindingResult().getAllErrors().get(0).getDefaultMessage()));
+    }
 
     @ExceptionHandler(JwtException.class)
     public ResponseEntity<ExceptionMessage> handle(JwtException e) {
@@ -32,5 +42,11 @@ public class TokenExceptionHandler {
     public ResponseEntity<ExceptionMessage> handle(TokenNotFoundException e) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                 .body(ExceptionMessage.of(StatusEnum.TOKEN_EXPIRED, e.getMessage()));
+    }
+
+    @ExceptionHandler(EmptyResultDataAccessException.class)
+    public ResponseEntity<ExceptionMessage> handle(EmptyResultDataAccessException e) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(ExceptionMessage.of(StatusEnum.TOKEN_EXPIRED, "해당 아이디를 가진 유저가 없습니다. 아이디 값을 다시 한번 확인하세요."));
     }
 }
