+++ src/main/java/co/kr/jurumarble/exception/user/UserIllegalStateException.java
@@ -4,7 +4,6 @@ import co.kr.jurumarble.exception.StatusEnum;
 import co.kr.jurumarble.exception.common.CustomException;
 import lombok.Getter;
 
-@Getter
 public class UserIllegalStateException extends CustomException {
     private final StatusEnum status;
     private static final String message = "MBTI 수정 후 2개월 내에 수정할 수 없습니다.";
